feat: correcciones controladorpoliza udistrital/arka_cliente#875
