+++ helpers/utilsHelper/utilsHelper.helper.go
@@ -192,3 +192,16 @@ func FindIdInArray(idsList []*models.Elemento, id int) (i int) {
 	}
 	return -1
 }
+
+// removeDuplicateIds Remueve de un vector los enteros duplicados
+func RemoveDuplicateIds(addrs []int) []int {
+	result := make([]int, 0, len(addrs))
+	temp := map[int]struct{}{}
+	for _, item := range addrs {
+		if _, ok := temp[item]; !ok {
+			temp[item] = struct{}{}
+			result = append(result, item)
+		}
+	}
+	return result
+}
